chore: unified rudderstack components to analytics folder
