+++ packages/bot-web-ui/src/analytics/rudderstack-quick-strategy.ts
@@ -1,5 +1,5 @@
 import { Analytics, TEvents } from '@deriv-com/analytics';
-import { ACTION, form_name, type TFormStrategy, type TSelectedStrategy } from './constants';
+import { ACTION, form_name, type TFormStrategy } from './constants';
 import { getRsStrategyType, getSubpageName, getTradeParameterData } from './utils';
 
 export const rudderStackSendQsOpenEvent = ({ subform_source }: TEvents['ce_bot_form']) => {
@@ -56,24 +56,6 @@ export const rudderStackSendQsSelectedTabEvent = ({ quick_strategy_tab }: TEvent
         form_name,
         subform_name: 'quick_strategy',
         quick_strategy_tab,
-    });
-};
-
-export const rudderStackSendSelectQsStrategyGuideEvent = ({ selected_strategy }: TSelectedStrategy) => {
-    Analytics.trackEvent('ce_bot_form', {
-        action: ACTION.SELECT_QUICK_STRATEGY_GUIDE,
-        form_name,
-        subpage_name: getSubpageName(),
-        strategy_name: getRsStrategyType(selected_strategy),
-    });
-};
-
-export const rudderStackSendTutorialSearchEvent = ({ search_term }: TEvents['ce_bot_form']) => {
-    Analytics.trackEvent('ce_bot_form', {
-        action: 'search',
-        form_name: 'ce_bot_form',
-        subpage_name: getSubpageName(),
-        search_term,
     });
 };
