+++ packages/bot-web-ui/src/components/trade-animation/trade-animation.tsx
@@ -6,7 +6,7 @@ import { localize } from '@deriv/translations';
 import ContractResultOverlay from 'Components/contract-result-overlay';
 import { contract_stages } from 'Constants/contract-stage';
 import { useDBotStore } from 'Stores/useDBotStore';
-import { rudderStackSendRunBotEvent } from '../../pages/bot-builder/quick-strategy/analytics/rudderstack-quick-strategy';
+import { rudderStackSendRunBotEvent } from '../../analytics/rudderstack-quick-strategy';
 import CircularWrapper from './circular-wrapper';
 import ContractStageText from './contract-stage-text';
