+++ packages/bot-web-ui/src/pages/bot-builder/quick-strategy/quick-strategy.tsx
@@ -6,7 +6,7 @@ import { MobileFullPageModal, Modal } from '@deriv/components';
 import { observer, useStore } from '@deriv/stores';
 import { localize } from '@deriv/translations';
 import { useDBotStore } from 'Stores/useDBotStore';
-import { rudderStackSendQsCloseEvent } from './analytics/rudderstack-quick-strategy';
+import { rudderStackSendQsCloseEvent } from '../../../analytics/rudderstack-quick-strategy';
 import DesktopFormWrapper from './form-wrappers/desktop-form-wrapper';
 import MobileFormWrapper from './form-wrappers/mobile-form-wrapper';
 import LossThresholdWarningDialog from './parts/loss-threshold-warning-dialog';
