+++ packages/bot-web-ui/src/pages/bot-builder/toolbox/toolbox.tsx
@@ -3,8 +3,8 @@ import classNames from 'classnames';
 import { Icon, Text } from '@deriv/components';
 import { observer, useStore } from '@deriv/stores';
 import { localize } from '@deriv/translations';
+import { rudderStackSendQsOpenEvent } from '../../../analytics/rudderstack-quick-strategy';
 import { useDBotStore } from '../../../stores/useDBotStore';
-import { rudderStackSendQsOpenEvent } from '../quick-strategy/analytics/rudderstack-quick-strategy';
 import ToolbarButton from '../toolbar/toolbar-button';
 import SearchBox from './search-box';
 import { ToolboxItems } from './toolbox-items';
